docs: 新增区块-行政区划选择器
