+++ docs/blocks/administrative-select.md
@@ -11,7 +11,7 @@ nav:
 
 ### 介绍
 
-用于快速查找省市县位置并迅速定位到指定位置
+用于快速查找中国省/市/县行政区域并快速定位的控件，基于 Ant Design 中的 [Cascader](https://ant-design.antgroup.com/components/cascader-cn#api) 组件封装而成
 
 ### 代码演示
 
@@ -21,12 +21,10 @@ nav:
 
 ### API
 
-| 参数           | 说明             | 类型                               | 默认值 |
-| -------------- | ---------------- | ---------------------------------- | ------ |
-| value          | 指定选中项       | `string[]｜number[]`               | `--`   |
-| onChange       | 选择完成后的回调 | `(value, selectedOptions) => void` | `--`   |
-| enableBoundary | 是否显示边界     | `boolean`                          | `true` |
-| autoFit        | 是否平移         | `boolean`                          | `true` |
-| boundaryLayer  | 边界 layer 属性  | `Omit<LineLayerProps, 'source'>`   | `--`   |
+| 参数 | 说明 | 类型 | 默认值 |
+| --- | --- | --- | --- |
+| autoFit | 是否控制地图自动平移到选项对应行政区域 | `boolean` | `true` |
+| enableBoundary | 是否在地图上展示行政区域边界 | `boolean` | `true` |
+| boundaryLayer | 边界线图层属性，可参考 [LineLayerProps](https://larkmap.antv.antgroup.com/components/layers/base-layers/line-layer#api) | `Omit<LineLayerProps, 'source'>` | `--` |
 
 其他参数可以参照 [Ant Design 5.0 Cascader](https://ant-design.antgroup.com/components/cascader-cn#api)
